Fix false start (#1)
* Add test exposing problem handling false start

* Fix Decode() to handle test case in last commit

Decoder.Decode() now saves each byte read (except the initial start
byte) into a buffer (byte slice). Bytes are read from Decoder.br and
saved in the buffer until the size of the buffer is the same as the last
length byte read.

After the buffer contains enough bytes, the creation of a packet is
attempted. If there is a checksum failure, bytes are removed from the
buffer until a start byte is found, and if one isn't found before the
buffer is completely emptied, bytes are read from the source until a
start byte is found. Then decoding starts over again as described above.

This change allows Decode() to recover from a false start byte present
in the byte stream without dropping any messages.

* Add TestDecodeMultipleFalseStarts()

* Update Decoder.Decode()

Now if an unknown message ID is seen in the byte stream, the caller of
Decode() can call Decode() again in case the unkown message ID was due
to a corrupt byte stream and not due to a wrong dialect.

* Add tests to decode two messages in a row

* Simplify use of startFound flag